move big tile and small tile into components folder
